pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poly-csp"
version = "0.1.0"
description = "Deterministic construction of helically ordered chiral stationary phase polymers."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]

# Keep dependencies minimal here — environment.yml handles versions
dependencies = [
  "numpy",
  "scipy",
  "pydantic",
  "hydra-core",
  "omegaconf",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "black",
  "mypy",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
